Removed history for dummy users (Revert me 2)

+++ view/src/main/java/com/noqapp/mobile/view/listener/NoQAppInitializationCheckBean.java
@@ -143,8 +143,10 @@ public class NoQAppInitializationCheckBean {
 
     @PostConstruct
     public void cleanForDummyUser() {
-        queueMobileService.deleteReferenceTo("100000000025");
-        queueMobileService.deleteReferenceTo("100000000064");
+        if (buildEnvironment.equalsIgnoreCase("prod")) {
+            queueMobileService.deleteReferenceTo("100000000025");
+            queueMobileService.deleteReferenceTo("100000000064");
+        }
     }
 
     @PreDestroy
